requirements.txt
python-telegram-bot==21.9
pymongo==4.10.1
dnspython==2.7.0
flask==3.0.0